Add devcontainer.json for Node.js & TypeScript development
This commit adds the devcontainer.json file, which is used for configuring a development container for Node.js and TypeScript projects. The file includes options for specifying the container image, features to add, forwarding ports, post-create commands, customizations, and connecting as root. For more information on how to use this file, please refer to the README at: https://github.com/devcontainers/templates/tree/main/src/typescript-node